Thanks Tohara for the contributions

+++ fabui/application/modules/objectmanager/objectmanager.php
@@ -13,7 +13,9 @@ class Objectmanager extends Module {
         }
 
 		$this -> lang -> load($_SESSION['language']['name'], $_SESSION['language']['name']);
-		shell_exec('sudo rm -r download');
+		if (file_exists('download')) {
+			shell_exec('sudo rm -r download');
+		}
 
 	}
 
@@ -840,7 +842,7 @@ class Objectmanager extends Module {
 				$this -> merge($object, $file);
 				break;
 			case 'print' :
-				redirect('create?obj=' . $object . '&file=' . $file);
+				redirect('make/print?obj=' . $object . '&file=' . $file);
 				break;
 		}
